Move test to integration
